requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
google-generativeai>=0.7.0
openai>=1.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
httpx>=0.27.0
pygments>=2.18.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
